test: Cobertura de tese no UserProfile
